--- PasswordTextField/PasswordTextField.xcodeproj/xcuserdata/kelson99hartle.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "F761E52A-DEB2-4CCD-82B6-5A0DB78BAEF2"
-   type = "1"
-   version = "2.0">
-</Bucket>
