--- PasswordTextField/PasswordTextField/Assets.xcassets/eyes-closed.imageset/Contents.json
@@ -1,23 +1,0 @@
-{
-  "images" : [
-    {
-      "idiom" : "universal",
-      "filename" : "eyes-closed.png",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "filename" : "eyes-closed@2x.png",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "filename" : "eyes-closed@3x.png",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
