+++ PasswordTextField1/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
